REST API/working/EIA API/requirements.txt
pandas==1.3.5
python_dateutil==2.8.2
requests==2.26.0
